server/requirements.txt
openenv-core>=0.2.3
pydantic>=2.0
fastapi>=0.100
uvicorn>=0.20
httpx>=0.25
openai>=1.0
